+++ src/main/java/com/feng/hotel/request/CreateOrderRequest.java
@@ -2,13 +2,12 @@ package com.feng.hotel.request;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
+import lombok.Data;
 
-import java.math.BigDecimal;
-import java.util.List;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
-
-import lombok.Data;
+import java.math.BigDecimal;
+import java.util.List;
 
 /**
  * @author Administrator
@@ -24,11 +23,11 @@ public class CreateOrderRequest {
 
     @ApiModelProperty(value = "押金")
     @NotNull(message = "押金不能为空")
-    private BigDecimal mortgage;
+    private Integer mortgage;
 
     @ApiModelProperty(value = "总价(交的钱-押金)")
     @NotNull(message = "总价不能为空")
-    private BigDecimal totalPrice;
+    private Integer totalPrice;
 
     @ApiModelProperty(value = "订单号", hidden = true)
     private String orderNo;
